+++ nodalia-cards.manifest.js
@@ -1,8 +1,8 @@
 export default {
-  "pkgVersion": "1.0.3-alpha.5",
-  "contentSha256_12": "41e8385633f5",
+  "pkgVersion": "1.0.3-alpha.6",
+  "contentSha256_12": "1a5948a0b364",
   "file": "nodalia-cards.bundle.js"
 };
-export const pkgVersion = "1.0.3-alpha.5";
-export const contentSha256_12 = "41e8385633f5";
+export const pkgVersion = "1.0.3-alpha.6";
+export const contentSha256_12 = "1a5948a0b364";
 export const file = "nodalia-cards.bundle.js";